Quiz Template: The score gets added, when the quiz is restarted/played again.
The score(total correct, total wrong) adds up with the quiz taken previously, if the quiz is restarted/played again. Also the Unanswered gets wrong.
This issue is fixed now. I have just initialized the variables: totalCorrect and totalWrong in the QuizModel to '0', and called this function from ScoreActivity, whenever the quiz is restarted.