Spellings Puzzle: Issue with scorecard and restarting the Spellings Puzzle.
ScoreActivity was meant only for Quiz and not for Spellings puzzle. It
was not able to handle the scores of Spellings puzzle. The issue was
fixed. Now ScoreActivity can handle scores from both Quiz and Spellings
Puzzle.
Also, there was an issue with restarting the Spellings Puzzle, the
scores start accumulating and also the index of the question was not set
to '0' while restarting the puzzle.